backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
black==26.1.0
click==8.3.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.128.7
h11==0.16.0
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.5.1
pluggy==1.6.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pymongo==4.16.0
pytest==9.0.2
python-dotenv==1.2.1
python-jose[cryptography]==3.4.0
python-multipart==0.0.22
pytokens==0.4.1
ruff==0.15.0
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
